# apps/armi-runtime/src/armi_runtime/interfaces/creator_web_resources/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatorExportRequest": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "directory_name": {
            "pattern": "[A-Za-z0-9][A-Za-z0-9._-]{0,63}",
            "title": "Directory Name",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "directory_name"
        ],
        "title": "CreatorExportRequest",
        "type": "object"
      },
...
      "DataRightsOrderCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/DataRightsOrderDetailResponse"
            },
            "title": "Orders",
            "type": "array"
          },
          "projection_version": {
            "const": "data-rights-order-collection.v3",
            "title": "Projection Version",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "orders"
        ],
        "title": "DataRightsOrderCollectionResponse",
        "type": "object"
      },
      "DataRightsOrderDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "effective_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Effective At",
            "type": "string"
          },
          "execution_status": {
            "enum": [
              "pending",
              "executing",
              "completed",
              "partial"
            ],
            "title": "Execution Status",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DataRightsOrderItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "newly_created": {
            "title": "Newly Created",
            "type": "boolean"
          },
          "order_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Order Id",
            "type": "string"
          },
          "order_kind": {
            "enum": [
              "stop_contact",
              "stop_use",
              "delete_related"
            ],
            "title": "Order Kind",
            "type": "string"
          },
          "projection_version": {
            "const": "data-rights-order-detail.v3",
            "title": "Projection Version",
            "type": "string"
          },
          "request_digest": {
            "pattern": "sha256:[0-9a-f]{64}",
            "title": "Request Digest",
            "type": "string"
          },
          "requester_kind": {
            "enum": [
              "creator",
              "other_human"
            ],
            "title": "Requester Kind",
            "type": "string"
          },
          "requester_party_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Requester Party Id",
            "type": "string"
          },
          "retention_reasons": {
            "items": {
              "enum": [
                "rights_enforcement",
                "shared_reference",
                "objective_history",
                "subject_continuity",
                "operator_managed_snapshot"
              ],
              "type": "string"
            },
            "maxItems": 5,
            "title": "Retention Reasons",
            "type": "array"
          },
          "scope_kind": {
            "enum": [
              "party_contact",
              "party_local_data"
            ],
            "title": "Scope Kind",
            "type": "string"
          },
          "scope_party_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Scope Party Id",
            "type": "string"
          },
          "status": {
            "const": "effective",
            "title": "Status",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/DataRightsTimelineItemResponse"
            },
            "minItems": 1,
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "order_id",
          "requester_party_id",
          "requester_kind",
          "order_kind",
          "scope_kind",
          "scope_party_id",
          "status",
          "execution_status",
          "request_digest",
          "effective_at",
          "completed_at",
          "newly_created",
          "items",
          "timeline",
          "retention_reasons"
        ],
        "title": "DataRightsOrderDetailResponse",
        "type": "object"
      },
...
      "DataRightsOrderRequest": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "order_kind": {
            "enum": [
              "stop_contact",
              "stop_use",
              "delete_related"
            ],
            "title": "Order Kind",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "order_kind"
        ],
        "title": "DataRightsOrderRequest",
        "type": "object"
      },
      "DataRightsOrderResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "effective_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Effective At",
            "type": "string"
          },
          "execution_status": {
            "enum": [
              "pending",
              "executing",
              "completed",
              "partial"
            ],
            "title": "Execution Status",
            "type": "string"
          },
          "newly_created": {
            "title": "Newly Created",
            "type": "boolean"
          },
          "order_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Order Id",
            "type": "string"
          },
          "order_kind": {
            "enum": [
              "stop_contact",
              "stop_use",
              "delete_related"
            ],
            "title": "Order Kind",
            "type": "string"
          },
          "projection_version": {
            "const": "data-rights-order-summary.v3",
            "title": "Projection Version",
            "type": "string"
          },
          "request_digest": {
            "pattern": "sha256:[0-9a-f]{64}",
            "title": "Request Digest",
            "type": "string"
          },
          "requester_kind": {
            "enum": [
              "creator",
              "other_human"
            ],
            "title": "Requester Kind",
            "type": "string"
          },
          "requester_party_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Requester Party Id",
            "type": "string"
          },
          "scope_kind": {
            "enum": [
              "party_contact",
              "party_local_data"
            ],
            "title": "Scope Kind",
            "type": "string"
          },
          "scope_party_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Scope Party Id",
            "type": "string"
          },
          "status": {
            "const": "effective",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "order_id",
          "requester_party_id",
          "requester_kind",
          "order_kind",
          "scope_kind",
          "scope_party_id",
          "status",
          "execution_status",
          "request_digest",
          "effective_at",
          "completed_at",
          "newly_created"
        ],
        "title": "DataRightsOrderResponse",
        "type": "object"
      },
      "DataRightsTimelineItemResponse": {
        "additionalProperties": false,
        "properties": {
          "event_kind": {
            "enum": [
              "order_effective",
              "item_status"
            ],
            "title": "Event Kind",
            "type": "string"
          },
          "item_id": {
            "anyOf": [
              {
                "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "status": {
            "enum": [
              "effective",
              "pending",
              "completed",
              "partial",
              "too_late",
              "unknown"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "event_kind",
          "occurred_at",
          "item_id",
          "status"
        ],
        "title": "DataRightsTimelineItemResponse",
        "type": "object"
      },
...
      "ErrorDescriptorResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategoryValue"
          },
          "code": {
            "pattern": "[A-Z][A-Z0-9_]{2,127}",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error_instance_id": {
            "anyOf": [
              {
                "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Instance Id"
          }
        },
        "required": [
          "category",
          "code"
        ],
        "title": "ErrorDescriptorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "RejectedOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptorResponse"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "status": {
            "const": "rejected",
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "pattern": "[0-9a-f]{32}",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "trace_id",
          "occurred_at",
          "message",
          "status",
          "error"
        ],
        "title": "RejectedOutcomeResponse",
        "type": "object"
      },
...
      "UnavailableOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptorResponse"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "recovery_hint": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Hint"
          },
          "status": {
            "const": "unavailable",
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "pattern": "[0-9a-f]{32}",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "trace_id",
          "occurred_at",
          "message",
          "status",
          "error"
        ],
        "title": "UnavailableOutcomeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "browserSessionBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ARMI Creator Interface",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/data-rights/orders": {
      "get": {
        "operationId": "listDataRightsOrders",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsOrderCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "List Creator Data Rights Orders"
      },
      "post": {
        "operationId": "createDataRightsOrder",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataRightsOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsOrderResponse"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsOrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Create Creator Data Rights Order"
      }
    },
    "/v1/data-rights/orders/{order_id}": {
      "get": {
        "operationId": "getDataRightsOrder",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsOrderDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Get Creator Data Rights Order"
      }
    },
    "/v1/data-rights/orders/{order_id}/retry": {
      "post": {
        "operationId": "retryDataRightsOrder",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsOrderResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Retry Creator Data Rights Order"
      }
    },
...
    "/v1/exports": {
      "post": {
        "operationId": "createCreatorExport",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatorExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorExportResponse"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Create Creator Export"
      }
    },
    "/v1/exports/{export_id}": {
      "get": {
        "operationId": "getCreatorExport",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Get Creator Export"
      }
    },
...
  }
}